requirements.txt
functions-framework==3.3.0
google-events==0.7.0
google-api-core==2.11.1
protobuf==4.23.3
google-cloud-firestore==2.11.1
cloudevents==1.9.0
requests==2.30.0
httpx==0.24.1